# Use a binary search tree