… src/lib/binarySearch.ts
… src/test/fakeIndexedDB/fakeIndexedDB.ts