Fixes #44

This replaces the current array-based approach with a simple binary search tree. The runtime of the benchmark is ~13x faster (using `hyperfine`):

- Before: **4.300** s ±  0.178 s
- After: **323.2 ms** ±  42.1 ms

My main goal with the binary search tree was simplicity of implementation. I figure `fake-indexeddb` has more value as a "reference implementation" for IndexedDB rather than a hyper-optimized database. (But it's still nice for it to be ~13x faster!)

I opted for a basic [red-black tree](https://en.wikipedia.org/wiki/Red%E2%80%93black_tree) with a deletion strategy borrowed from the [scapegoat tree](https://en.wikipedia.org/wiki/Scapegoat_tree). Deletion in a scapegoat tree is much simpler: you just mark nodes with `deleted` tombstones and rebuild the tree if there are too many tombstones, which avoids rotations to rebalance. I felt this was a good compromise to try to keep things simple. I doubt most `fake-indexeddb` users are deleting a bunch of nodes (you'd more normally just throw away the whole database between each test), and even if they do, the amortized cost of the deletions is still `O(log(n))`.

(Using a scapegoat tree itself is a nice idea, but it falls apart on insertions. The `FDBObjectStore.count performance should be reasonable` test is extremely slow with a scapegoat tree, because nodes are inserted in sorted order, which basically devolves into a linked list where the tree is rebuilt on every insertion. Whereas with the current tree, that particular test is only slightly slower: 214ms vs 127ms. It could probably be faster if I avoided recursion/generators, but again I was opting for readability.)