@nolanlawson 2025-08-27
^ Made some optimizations, the `FDBObjectStore.count performance should be reasonable` test is now down to ~162ms.